test: split models.spec
